# swagger.yaml
openapi: 3.0.0
info:
  title: Library API
  version: 1.0.0
  description: API for managing books and authors

servers:
  - url: http://localhost:3000

tags:
  - name: Authors
    description: Operations related to authors
  - name: Books
    description: Operations related to books

paths:
  /api/authors:
    get:
      tags: [Authors]
      summary: Get all authors
      responses:
        '200':
          description: List of authors

    post:
      tags: [Authors]
      summary: Create one or more authors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Author'
                - type: array
                  items:
                    $ref: '#/components/schemas/Author'
      responses:
        '201':
          description: Author(s) created
        '422':
          description: Validation error

  /api/authors/{id}:
    get:
      tags: [Authors]
      summary: Get author by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Author found
        '404':
          description: Author not found

    put:
      tags: [Authors]
      summary: Update author by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        '200':
          description: Author updated
        '422':
          description: Validation error
        '404':
          description: Author not found

    delete:
      tags: [Authors]
      summary: Delete author by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Author deleted
        '404':
          description: Author not found

  /api/books:
    get:
      tags: [Books]
      summary: Get all books
      responses:
        '200':
          description: List of books

    post:
      tags: [Books]
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Book created
        '422':
          description: Validation error

  /api/books/{id}:
    get:
      tags: [Books]
      summary: Get book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book found
        '404':
          description: Book not found

    put:
      tags: [Books]
      summary: Update book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book updated
        '422':
          description: Validation error
        '404':
          description: Book not found

    delete:
      tags: [Books]
      summary: Delete book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book deleted
        '404':
          description: Book not found

components:
  schemas:
    Author:
      type: object
      required:
        - name
        - bio
        - birthdate
      properties:
        name:
          type: string
        bio:
          type: string
        birthdate:
          type: string
          format: date

    Book:
      type: object
      required:
        - title
        - author
      properties:
        title:
          type: string
        author:
          type: string